requirements.txt
load_dotenv==0.1.0
djangorestframework==3.15.1
djangorestframework-api-key==3.0.0
requests==2.32.3
psycopg==3.1.19
psycopg2==2.9.9
gunicorn==22.0.0
whitenoise==6.6.0
pytest==8.2.2
pytest-django==4.8.0